# KNOX-2136 - Caching credentials in DefaultKeystoreService when an alias is being added or loaded from keystore and using a different cache implementation