--- a/pom.xml
+++ b/pom.xml
@@ -154,6 +154,7 @@
         <asm.version>7.2</asm.version>
         <aspectj.version>1.9.4</aspectj.version>
         <bcprov-jdk15on.version>1.64</bcprov-jdk15on.version>
+        <ben-manes.caffeine.version>2.8.0</ben-manes.caffeine.version>
         <buildnumber-maven-plugin.version>1.4</buildnumber-maven-plugin.version>
         <cglib.version>3.3.0</cglib.version>
         <checkstyle.version>8.26</checkstyle.version>
@@ -2202,6 +2203,21 @@
                 <groupId>com.cloudera.api.swagger</groupId>
                 <artifactId>cloudera-manager-api-swagger</artifactId>
                 <version>${cloudera-manager-api-swagger.version}</version>  <!-- or CM version 6.0 and above -->
+            </dependency>
+            <dependency>
+                <groupId>com.github.ben-manes.caffeine</groupId>
+                <artifactId>caffeine</artifactId>
+                <version>${ben-manes.caffeine.version}</version>
+                <exclusions>
+                    <exclusion>
+                        <groupId>org.checkerframework</groupId>
+                        <artifactId>checker-qual</artifactId>
+                    </exclusion>
+                    <exclusion>
+                        <groupId>com.google.errorprone</groupId>
+                        <artifactId>error_prone_annotations</artifactId>
+                    </exclusion>
+                </exclusions>
             </dependency>
 
             <!-- ********** ********** ********** ********** ********** ********** -->